test: migrate document_service_status SQL tests to testcontainers
